+++ src/Layout/CardButton/CardButton.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import BotaoRosaPrincipal from "../../components/PinkButton";
-import Mulher from "../../Imagens/Mulher.png"
+import Mulher from "../../assets/mulher.png"
 import "./CardButton.css";
 
 const CardButton = (props) => {
